defined data graph model
